+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Changes to be tagged
 ### Added
 - `svtBuffTurnExtend` to nice servant script.
+- `NiceConstantStr` export file.
 
 ## 5.76.0 - 2022-01-23
 ### Added